requirements.txt
# Core dependencies
manim>=0.18.0
openai>=1.3.0
gradio>=4.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Text and LaTeX rendering
manimpango>=0.5.0  # For text rendering in Manim
latex>=0.7.0  # Python LaTeX utilities

# Note: The following LaTeX distributions need to be installed via system package manager:
# Windows: MiKTeX (https://miktex.org/download)
# Ubuntu/Debian: texlive-full
# macOS: MacTeX

# Optional but recommended for development
ipython>=8.0.0  # For interactive development
pytest>=7.0.0  # For testing
black>=22.0.0  # For code formatting

# Additional System Requirements:
# FFmpeg - Required for video rendering with Manim
# Installation options:
# 1. Windows: Download from https://www.gyan.dev/ffmpeg/builds/
#    - Recommended: Get the "ffmpeg-git-full.7z" release
#    - Extract and add the bin folder to system PATH
# 2. Alternative: Install via package managers
#    - Windows (Chocolatey): choco install ffmpeg
#    - Linux: apt-get install ffmpeg
#    - macOS: brew install ffmpeg
